refactor: break test_metrics file into one file for each metric
